Extract application link to its own file
This was causing a dependency cycle
