+++ server/utils/placementRequests/index.ts
@@ -4,4 +4,6 @@ export * as tableUtils from './table'
 export { assessmentSummary } from './assessmentSummaryList'
 export { matchingInformationSummary } from './matchingInformationSummaryList'
 export { documentSummary } from './documentSummaryList'
-export { formatReleaseType, searchButton, applicationLink, assessmentLink } from './utils'
+export { formatReleaseType, searchButton, assessmentLink } from './utils'
+export { mapPageForSummaryList, pagesForReview } from './checkYourAnswersUtils'
+export { applicationLink } from './applicationLink'
