+++ server/utils/placementRequests/utils.ts
@@ -1,11 +1,10 @@
-import { PlacementApplication, PlacementRequest } from '@approved-premises/api'
+import { PlacementRequest } from '@approved-premises/api'
 import { BedSearchParametersUi } from '@approved-premises/ui'
 import { allReleaseTypes } from '../applications/releaseTypeUtils'
 import { linkTo } from '../utils'
 
 import paths from '../../paths/match'
 import assessPaths from '../../paths/assess'
-import applyPaths from '../../paths/apply'
 
 export const mapPlacementRequestToBedSearchParams = ({
   duration,
@@ -36,12 +35,6 @@ export const searchButton = (placementRequest: PlacementRequest) =>
     { id: placementRequest.id },
     { text: 'Search', attributes: { class: 'govuk-button' } },
   )
-
-export const applicationLink = (
-  placementRequestOrApplication: PlacementRequest | PlacementApplication,
-  text: string,
-  hiddenText: string,
-) => linkTo(applyPaths.applications.show, { id: placementRequestOrApplication.applicationId }, { text, hiddenText })
 
 export const assessmentLink = (placementRequest: PlacementRequest, text: string, hiddenText: string) =>
   linkTo(assessPaths.assessments.show, { id: placementRequest.assessmentId }, { text, hiddenText })
